server/requirements.txt
# memOS Server Requirements for Recovery Bot Integration
# Core Memory Framework
mem0ai>=0.1.0                    # Primary memory framework
chromadb>=0.4.18                 # Vector database backend
sentence-transformers>=2.2.2     # Text embeddings

# FastAPI and Web Framework
fastapi>=0.104.1                 # Web framework
uvicorn>=0.24.0                  # ASGI server
pydantic>=2.5.0                  # Data validation
python-multipart>=0.0.6         # Form data support

# Database Integration
psycopg2-binary>=2.9.7           # PostgreSQL adapter
pgvector>=0.2.3                  # Vector extension for PostgreSQL
sqlalchemy>=2.0.23               # ORM
alembic>=1.13.1                  # Database migrations

# Authentication and Security
PyJWT>=2.8.0                     # JWT tokens
argon2-cffi>=23.1.0              # Password hashing
cryptography>=41.0.7             # Encryption utilities
python-jose[cryptography]>=3.3.0 # JWT handling

# Redis and Caching
redis>=5.0.1                     # Redis client
aioredis>=2.0.1                  # Async Redis client

# LLM Integration
httpx>=0.25.2                    # HTTP client for Ollama
openai>=1.3.8                    # OpenAI-compatible client
ollama>=0.1.7                    # Ollama Python client

# Data Processing
numpy>=1.24.4                    # Numerical operations
pandas>=2.1.3                    # Data manipulation (optional)

# Monitoring and Logging
structlog>=23.2.0                # Structured logging
prometheus-client>=0.19.0        # Metrics
python-json-logger>=2.0.7        # JSON logging

# Development and Testing
pytest>=7.4.3                    # Testing framework
pytest-asyncio>=0.21.1           # Async testing
black>=23.11.0                   # Code formatting
ruff>=0.1.6                      # Linting

# Environment Management
python-dotenv>=1.0.0             # Environment variables
pyyaml>=6.0.1                    # YAML configuration

# HIPAA Compliance and Privacy
anonymizer>=0.1.0                # Data anonymization
audit-logger>=1.0.0              # Audit trail logging

# Text-to-Speech
edge-tts>=6.1.0                  # Microsoft Edge TTS (free, high-quality voices)
pydub>=0.25.1                    # Audio format conversion (MP3 to WAV for Android)